BAU: Stop kong re-downloading deb on every run
At present, the deb download occasionally fails, stopping the infra
pipeline.

The debs are downloaded to /var/tmp and have the patch version in the
filename, so there is no need to re-download the deb on each run.

By adding 'force=no', we stop get_url from downloading if the file
already exists at destination - see
http://docs.ansible.com/ansible/latest/get_url_module.html and
https://github.com/ansible/ansible/issues/2416 for detail